Fix BossEventSerializer_v291 deserialization (#198)

+++ bedrock-codec/src/main/java/org/cloudburstmc/protocol/bedrock/codec/v291/serializer/BossEventSerializer_v291.java
@@ -21,7 +21,7 @@ public class BossEventSerializer_v291 implements BedrockPacketSerializer<BossEve
 
     @Override
     public void deserialize(ByteBuf buffer, BedrockCodecHelper helper, BossEventPacket packet) {
-        packet.setBossUniqueEntityId(VarInts.readInt(buffer));
+        packet.setBossUniqueEntityId(VarInts.readLong(buffer));
         packet.setAction(BossEventPacket.Action.values()[VarInts.readUnsignedInt(buffer)]);
         this.deserializeAction(buffer, helper, packet);
     }
